third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/gen-get-same-object.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('TwoHeartRateServicesAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['generic_access']}))
    .then(device => device.gatt.connect())
    .then(gattServer => Promise.all([
      gattServer.getPrimaryService('heart_rate'),
      gattServer.getPrimaryService('heart_rate')]))
    .then(services_arrays => {
      // Convert to arrays if necessary.
      for (let i = 0; i < services_arrays.length; i++) {
        services_arrays[i] = [].concat(services_arrays[i]);
      }

      for (let i = 1; i < services_arrays.length; i++) {
        assert_equals(services_arrays[0].length, services_arrays[i].length);
      }

      let base_set = new Set(services_arrays[0]);
      for (let services of services_arrays) {
        services.forEach(service => assert_true(base_set.has(service)));
      }
    });
}, 'Calls to getPrimaryService should return the same object.');

</script>
